feat: add wallet-backed instruction items with Brick Money minor-unit pricing

+++ database/migrations/2024_08_02_317537_create_instruction_items_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('index')->unique();
             $table->string('type')->index();
-            $table->decimal('price', 12, 2)->default(0);
+            $table->bigInteger('price')->default(0); // minor units
             $table->string('currency', 3)->default('PHP');
             $table->json('meta')->nullable();
             $table->nullableMorphs('revenue_destination');
